+++ e2e_test/lib/nativewrapper.dart
@@ -5,11 +5,15 @@ doSomething() {
   String str = "ciao";
   int z = str.codeUnitAt(1);
 
-
   double w = 5.5.floorToDouble();
 
   String rev = "ciao".toUpperCase();
 
   List<int> codes = rev.codeUnits;
 
+  // Some static methods :
+
+  int n = int.parse("5");
+  double d = double.parse("5.4");
+  num N = num.parse("3.14");
 }